+++ libnds-sys/src/arm7_error.rs
@@ -1,0 +1,1 @@
+compile_error!("These registers can only be used when compiling for ARM7");
