+++ libnds-sys/src/atomics.rs
@@ -3,6 +3,7 @@ use super::arm7_bindings::*;
 #[cfg(target_os = "nintendo_ds_arm9")]
 use super::arm9_bindings::*;
 
+
 use libc::c_int;
 use libc::c_void;
