+++ libnds-sys/src/bios_registers.rs
